savings_champion/requirements.txt
lxml
nose
Django<1.8
pillow
django-piston
django-extensions
django-tinymce
sorl-thumbnail
suds
feedparser
django-registration-redux
python-dateutil
python-memcached

#  Waiting for update to Django>1.7
django-pipeline<1.6.0

requests

django-pipeline-forgiving
django-grappelli
django-ckeditor<4.4.8
django-filebrowser
raven
celery[redis]
uwsgi
psycopg2
django-select2
django-crispy-forms
pyotp
django-celery
django-icybackup
newrelic
djangorestframework
djangorestframework-xml
sympy
django-nose
mock
parsedatetime
simple-salesforce
git+git://github.com/savingschampion/django-allauth.git@master#egg=django-allauth
apache-libcloud
django-storages-redux
boto
Collectfast
sortedcontainers
django-debug-toolbar
django-debug-toolbar-line-profiler
redis
hiredis
msgpack-python
behave
django-behave
django-waffle
universal-analytics-python
django-guardian
pysendy
django-report-builder
django-markupfield
Markdown
django-cors-headers
python-statsd
arrow
django-redis
mailchimp3
